requirements.txt
jinja2
pytest
graphql-core
